Added our MySQL Client handler and preforker.
This is so we can prefork mysql clients for the current workload, but we
also have a max client limit because we dont need to have one client for
each query that will just overload the server plus connecting to a mysql
socket and initializing a client utilizes too much cpu to be optimal for
each query. Also there goes some good caching. So we limit to a certain
amount and try to keep the connections alive.